+++ NEWS.md
@@ -1,3 +1,9 @@
-# ssa.hts (development version)
+# ssa.hts 0.1.0
 
 * Initial CRAN submission.
+
+* Can now create an hankel matrix from a time series.
+* Can now create a trajectory matrix from a time series.
+
+
+